[10/n bindings refactor] the _king is dead, long live the king
Relocates the small amount of code that actually lived in monarch._monarch, and removes it.

Differential Revision: [D75027282](https://our.internmc.facebook.com/intern/diff/D75027282/)

[ghstack-poisoned]
