--- python/monarch/_monarch/__init__.py
@@ -1,3 +1,0 @@
-# Import before monarch to pre-load torch DSOs as, in exploded wheel flows,
-# our RPATHs won't correctly find them.
-import torch  # isort:skip
